requirements.txt
transformers
datasets
trl
accelerate
bitsandbytes
torch
pyyaml